+++ sunpy/net/tests/test_scraper.py
@@ -2,19 +2,11 @@ import datetime
 from unittest.mock import Mock, patch
 
 import pytest
-from dateutil.relativedelta import relativedelta
 
 from sunpy.data.test import rootdir
 from sunpy.extern import parse
-from sunpy.net.scraper import Scraper, get_timerange_from_exdict
+from sunpy.net.scraper import Scraper
 from sunpy.time import TimeRange, parse_time
-
-TIMEPATTERN_EXAMPLES = [
-    ('%b%y', relativedelta(months=1)),
-    ('%m%y', relativedelta(months=1)),
-    ('%H%d', relativedelta(hours=1)),
-    ('%y%b', relativedelta(months=1)),
-]
 
 
 def testDirectoryDatePattern():
@@ -67,12 +59,6 @@ def testNoDateDirectory():
     directory_list = ['mySpacecraft/myInstrument/xMinutes/']
     timerange = TimeRange('2009/11/20', '2010/01/03')
     assert s.range(timerange) == directory_list
-
-
-@pytest.mark.parametrize(('pattern', 'mintime'), TIMEPATTERN_EXAMPLES)
-def test_smallerPattern(pattern, mintime):
-    assert mintime == Scraper('')._smallerPattern(pattern)
-
 
 def testDirectoryRangeHours():
     s = Scraper('{{year:4d}}{{month:2d}}{{day:2d}}_{{hour:2d}}/{{hour:2d}}{{minute:2d}}.csv')
@@ -232,19 +218,6 @@ def test_extract_files_meta():
     assert metalist1[3]['CAR_ROT'] == 2226
     assert metalist1[-1]['url'] == urls[-1]
 
-
-@pytest.mark.parametrize(('exdict', 'start', 'end'), [
-    ({"year": 2000}, '2000-01-01 00:00:00', '2000-12-31 23:59:59.999000'),
-    ({"year": 2016, "month": 2}, '2016-02-01 00:00:00', '2016-02-29 23:59:59.999000'),
-    ({'year': 2019, 'month': 2, 'day': 28}, '2019-02-28 00:00:00', '2019-02-28 23:59:59.999000'),
-    ({'year': 2020, 'month': 7, 'day': 31, 'hour': 23, 'minute': 59, 'second': 59},
-     '2020-07-31 23:59:59', '2020-07-31 23:59:59.999000')])
-def test_get_timerange_with_extractor(exdict, start, end):
-    tr = TimeRange(start, end)
-    file_timerange = get_timerange_from_exdict(exdict)
-    assert file_timerange == tr
-
-
 @pytest.mark.remote_data
 def test_yearly_overlap():
     # Check that a time range that falls within the interval that a file represents
@@ -255,33 +228,3 @@ def test_yearly_overlap():
     # Should return a single file for 2013
     trange = TimeRange("2013-01-02", "2013-01-03")
     assert len(scraper.filelist(trange)) == 1
-
-
-def test_check_timerange():
-    s0 = Scraper('{{year:4d}}.fits')
-    # Valid time range for 2014.fits is the whole of 2014
-    # Test different cases to make sure check_timerange is working as expected
-
-    # Interval exactly on lower boundary
-    assert s0._check_timerange('2014.fits', TimeRange("2013-06-01", "2014-01-01"))
-    # Overlaps lower boundary
-    assert s0._check_timerange('2014.fits', TimeRange("2013-06-01", "2014-01-02"))
-    # Overlaps upper and lower boundary
-    assert s0._check_timerange('2014.fits', TimeRange("2013-06-01", "2015-01-02"))
-    # Entirely within both boundaries
-    assert s0._check_timerange('2014.fits', TimeRange("2014-06-01", "2014-07-02"))
-    # Overlaps upper boundary
-    assert s0._check_timerange('2014.fits', TimeRange("2014-06-01", "2015-01-02"))
-    # Interval exactly on upper boundary
-    assert not s0._check_timerange('2014.fits', TimeRange("2015-01-01", "2015-01-02"))
-
-    # Interval below both boundaries
-    assert not s0._check_timerange('2014.fits', TimeRange("2002-01-01", "2013-01-02"))
-    # Interval above both boundaries
-    assert not s0._check_timerange('2014.fits', TimeRange("2022-01-01", "2025-01-02"))
-
-    s1 = Scraper('{{year:4d}}-{{month_name:l}}.fits')
-    assert s1._check_timerange('2014-March.fits', TimeRange("2014-03-01", "2004-04-01"))
-
-    s2 = Scraper('{{year:4d}}-{{month_name_abbr:l}}-{{day:2d}}.fits')
-    assert s2._check_timerange('2004-Mar-06.fits', TimeRange("2004-03-06", "2004-03-07"))
